+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.3.0] - Unreleased
+* Deprecated `findOneByIri`, `findOneBy`, `findBy`, `findAll` methods in `AbstractMicroserviceHttpRepository`
+* Added `fetchOneByIri`, `fetchOneBy`, `fetchBy`, `fetchAll` methods in `AbstractMicroserviceHttpRepository`
 * Deprecated `PaginatedCollectionIterator::iterateOver`
 * Added `PaginatedCollectionIterator::iterateItems`
 * Added `PaginatedCollectionIterator::iteratePages`